+++ mockingbird/src/commonMain/kotlin/com/careem/mockingbird/util/getSystemTimeInMillis.kt
@@ -1,0 +1,3 @@
+package com.careem.mockingbird.util
+
+internal expect fun getSystemTimeInMillis(): Long