+++ mockingbird/src/jsMain/kotlin/com/careem/mockingbird/test/sleep.kt
@@ -1,0 +1,9 @@
+package com.careem.mockingbird.test
+
+import com.careem.mockingbird.util.getSystemTimeInMillis
+
+actual fun sleep(millis: Long) { // TODO find a better way for this that doesn't keep thread busy
+    val end = getSystemTimeInMillis() + millis
+    while (getSystemTimeInMillis() < end) {
+    }
+}