+++ mockingbird/src/jvmMain/kotlin/com/careem/mockingbird/util/getSystemTimeInMillis.kt
@@ -1,0 +1,3 @@
+package com.careem.mockingbird.util
+
+internal actual fun getSystemTimeInMillis() = System.currentTimeMillis()